// src/app/models/passengers.modals.ts
export class Passengers {
  'id': number;
  'name': string;
  'passport': string;
  'address': string;
  'seatType': string;
  'seatNumber': string;
  'flightNumber': string;
  'contactNumber': string;
  'dateOfBirth': string;
  'ancillaryServices': Array<string>;
  'specialMeals': boolean;
  'shoppingItems': Array<string>;
  'checkedIn': boolean;
  'infant': boolean;
  'wheelChair': boolean;
}